requirements.txt
Flask
flask-cors
pymongo
openpyxl
python-dotenv 